Refactor DTOs: Merge DataTypes into Models.DTO and remove project
