+++ src/Argon.Api/Argon.Api.csproj
@@ -45,7 +45,6 @@
 
     <ItemGroup>
         <ProjectReference Include="..\Argon.Contracts\Argon.Contracts.csproj"/>
-        <ProjectReference Include="..\DataTypes\DataTypes.csproj"/>
         <ProjectReference Include="..\Grains.Interface\Grains.Interface.csproj" />
         <ProjectReference Include="..\ServiceDefaults\ServiceDefaults.csproj"/>
     </ItemGroup>
