+++ src/Argon.Api/Features/Sfu/ArgonSelectiveForwardingUnit.cs
@@ -1,12 +1,12 @@
 namespace Argon.Api.Features.Sfu;
 
 using System.IdentityModel.Tokens.Jwt;
-using DataTypes;
 using Flurl.Http;
 using LiveKit.Proto;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
+using Models.DTO;
 
 #if DEBUG
 public class ArgonSfuTestController : ControllerBase
