+++ src/Argon.Api/Features/Sfu/SfuPermissionFlags.cs
@@ -1,8 +1,8 @@
 ﻿namespace Argon.Api.Features.Sfu;
 
-using DataTypes;
 using LiveKit.Proto;
 using Microsoft.OpenApi.Extensions;
+using Models.DTO;
 
 [Flags]
 public enum SfuPermissionFlags
