+++ src/Argon.Api/Grains/UserManager.cs
@@ -4,6 +4,7 @@ using Entities;
 using global::Grains.Interface;
 using Microsoft.EntityFrameworkCore;
 using Models;
+using Models.DTO;
 using Services;
 
 public class UserManager(IPasswordHashingService passwordHashingService, ApplicationDbContext context) : Grain, IUserManager
