--- src/DataTypes/DataTypes.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Argon.Sfu.Protocol" Version="1.26.0"/>
-        <PackageReference Include="MemoryPack.Core" Version="1.21.3"/>
-        <PackageReference Include="Microsoft.Orleans.Serialization.Abstractions" Version="8.2.0"/>
-    </ItemGroup>
-
-</Project>
