+++ src/Grains.Interface/IServerManager.cs
@@ -1,9 +1,7 @@
 namespace Grains.Interface;
 
-using System.Runtime.Serialization;
-using MemoryPack;
-using MessagePack;
 using Models;
+using Models.DTO;
 using Orleans;
 
 public interface IServerManager : IGrainWithGuidKey
@@ -20,12 +18,3 @@ public interface IServerManager : IGrainWithGuidKey
     [Alias(nameof(DeleteServer))]
     Task DeleteServer();
 }
-
-[DataContract, MemoryPackable(GenerateType.VersionTolerant), MessagePackObject, Serializable, GenerateSerializer, Alias(nameof(ServerInput))]
-public sealed record ServerInput(
-    [property: DataMember(Order = 0), MemoryPackOrder(0), Key(0), Id(0)]
-    string Name,
-    [property: DataMember(Order = 1), MemoryPackOrder(1), Key(1), Id(1)]
-    string? Description,
-    [property: DataMember(Order = 2), MemoryPackOrder(2), Key(2), Id(2)]
-    string? AvatarUrl);
