+++ src/Models/DTO/ArgonServerId.cs
@@ -1,0 +1,3 @@
+namespace Models.DTO;
+
+public record struct ArgonServerId(Guid id);
