+++ src/Models/DTO/ArgonUserId.cs
@@ -1,0 +1,6 @@
+namespace Models.DTO;
+
+public record struct ArgonUserId(Guid id)
+{
+    public string ToRawIdentity() => id.ToString("N");
+}
