+++ src/Models/DTO/EphemeralChannelInfo.cs
@@ -1,0 +1,5 @@
+namespace Models.DTO;
+
+using LiveKit.Proto;
+
+public record struct EphemeralChannelInfo(ArgonChannelId channelId, string sid, Room room);
