+++ src/Models/Models.csproj
@@ -7,6 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <PackageReference Include="Argon.Sfu.Protocol" Version="1.26.0" />
         <PackageReference Include="MemoryPack.Core" Version="1.21.3"/>
         <PackageReference Include="MessagePack.Annotations" Version="2.5.172"/>
         <PackageReference Include="Microsoft.Orleans.Serialization.Abstractions" Version="8.2.0"/>
